fix(files-provider): lint fixes
