+++ services/workflows-service/.eslintrc.cjs
@@ -38,6 +38,8 @@ module.exports = {
     '@typescript-eslint/no-non-null-assertion': 'off',
     '@typescript-eslint/no-unsafe-argument': 'off',
     '@typescript-eslint/no-unsafe-return': 'off',
+    '@typescript-eslint/no-unsafe-assignment': 'off',
+    '@typescript-eslint/ban-ts-comment': 'off',
   },
   ignorePatterns: ['.eslintrc.cjs'],
 };
