+++ services/workflows-service/src/providers/file/types/interfaces.ts
@@ -1,20 +1,5 @@
-import { HttpFileService } from '@/providers/file/file-provider/http-file.service';
-import { AwsS3FileService } from '@/providers/file/file-provider/aws-s3-file.service';
-import { LocalFileService } from '@/providers/file/file-provider/local-file.service';
 import { Readable } from 'stream';
-
-export type TRemoteUri = string;
-export type TLocalFilePath = string;
-export type TS3BucketConfig = {
-  bucketName: string;
-  fileNameInBucket: string; // id in bucket
-  private: boolean | true;
-  uri?: TRemoteUri;
-};
-
-export type TRemoteFileConfig = TS3BucketConfig | TRemoteUri;
-export type TStringInt = string | number;
-export type TFileServiceProvider = HttpFileService | AwsS3FileService | LocalFileService;
+import { TLocalFilePath, TRemoteFileConfig } from './files-types';
 
 export interface IFileProvider {
   isRemoteFileExists(remoteFileConfig: TRemoteFileConfig): Promise<boolean>;
